Build chai.js with webpack
Summary: By bundling our Chai dependency with `webpack` instead of `build-common`, we can create a single bundle without any `require` statements that Kake can analyze with `analyze_for_dependencies`.

Test Plan: It's sufficent to `make chai.js` in this repo and verify that a bundle is produced that has all the exports of Chai. I'll add more of a test plan to the corresponding change in `webapp`.

Reviewers: #webpack

Differential Revision: https://phabricator.khanacademy.org/D48699